requirements.txt
playwright>=1.40.0
python-dotenv>=1.0.0
fastapi>=0.108.0
uvicorn>=0.25.0
jinja2>=3.1.2
plotly>=5.18.0
pandas>=2.1.0
